Closes #1 Setup TypeScript for Front-End
